requirements.txt
py~=1.9.0
pip~=20.2.3
attrs~=20.2.0
toml~=0.10.1
Werkzeug~=1.0.1
click~=7.1.2
Jinja2~=2.11.2
itsdangerous~=1.1.0
Flask~=1.1.2
Flask-Login~=0.5.0
Flask-SQLAlchemy~=2.4.4
MarkupSafe~=1.1.1
pytest~=6.0.2
pluggy~=0.13.1
iniconfig~=1.0.1
pyparsing~=2.4.7
six~=1.15.0
setuptools~=50.3.0
SQLAlchemy~=1.3.19
packaging~=20.4
pyotp~=2.4.0
pyjwt~=1.0.0
cryptography~=2.9.2